third_party/blink/web_tests/fast/forms/input-valueasnumber-unsupported-expected.txt

CONSOLE WARNING: The specified value "0" does not conform to the required format.  The format is "#rrggbb" where rr, gg, bb are two-digit hexadecimal numbers.
Tests for .valueAsNumber with non-supported input types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check for type=button
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=checkbox
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=color
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=email
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=file
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=hidden
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=image
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=khtml_isindex
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=password
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=radio
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=reset
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=search
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=submit
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=tel
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=text
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
Check for type=url
PASS input.valueAsNumber is Number.NaN
PASS input.valueAsNumber = 0 threw exception InvalidStateError: Failed to set the 'valueAsNumber' property on 'HTMLInputElement': This input element does not support Number values..
PASS successfullyParsed is true

TEST COMPLETE
